pyproject.toml
[tool.poetry]
name = "rt-detr-barcode-detection"
version = "0.1.0"
description = ""
authors = ["vladimir-sashin <sashin.v.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
clearml = "^1.15.1"
jinja2 = "3.0.3"    # Specific version required by fiftyone
omegaconf = "^2.3.0"
pydantic = "^2.7.1"
onnxruntime = "^1.18.1"
fiftyone = "^0.24.1"
ipywidgets = "^8.1.3"   # Required by fiftyone to plot confusion matrix
kaleido = "0.1.0.post1"


[tool.poetry.group.CI.dependencies]
wemake-python-styleguide = "^0.19.2"
pre-commit = "^3.7.0"


[tool.poetry.group.rtdetr.dependencies]
torch = "2.0.1"
torchvision = "0.15.2"
onnx = "1.14.0"
pycocotools = "^2.0.7"
pyyaml = "^6.0.1"
scipy = "^1.13.0"
tensorboard = "^2.16.2"


[tool.poetry.group.data.dependencies]
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
gdown = "^5.1.0"


[tool.vulture]
min_confidence = 80
paths = ["source"]
ignore_names = ["args", "kwargs", "batch_idx"]
sort_by_size = true
verbose = false


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
